Created libGDX with IntelliJ plus most base files
Still nothing to show for the 3+ hours of work it took to reorganize all of this, however it should hopefully be 10000x more modular than before. Should definitely have the ability for literally anything to be a constant that can be changed in settings. Left off after completing Placement, need to complete Board, don't forget to test all methods, classes, etc. and confirm that all of it works.