Simple Input Handling + Menu + Centralization in AzulGameScreen
Probably want to move a fair bit of stuff out of AzulGameScreen and put it somewhere else, but for now it is working well.
Input listeners now need to be updated to actually do processing (such as taking tiles and actually making the game progress and whatnot)
Menu needs to look nicer and larger, also background texture needs to be played around with to allow for anchoring.
Need to fix zoom for camera motion (slightly)
Need to add a clickable way of exiting out of a target (as in if you just click in the background, you should reset your target).
As always, check TODOs!